CONVERSION RATES UPDATE:
Developed a new api route that handles the currency conversion for available rates.
Implemented a useEffect that calls the conversion api route passing in the FROM & TO currency along with the amount to be converted.
Implemented a amount change function that handles the user amount input.
Set a condition statement that accommodate for the currency conversion from both the FROM & TO currency box to enable dynamic conversion from both input boxes.